cox bid adapter converted to module
